requirements.txt
jsonschema==3.2.0
docopt==0.6.2
yapf==0.30.0
kafka-python==2.0.1
psycopg2==2.8.6
requests==2.24.0
testing.postgresql==1.3.0
